[Math] Separate out util math from regular math
Topic: math-splitup
Relative:
Reviewers:
